chat: Move ChatMessages+Input layout component to components folder.
