+++ src/components/Chat/ChatMessages.css
@@ -1,0 +1,11 @@
+.ChatMessages {
+  height: 100%;
+  overflow-y: scroll;
+  font-size: 10pt;
+  /* Fix for tooltips close to the edge. The tooltips themselves work fine,
+   * but the element that's used to position the tooltips correctly can overflow
+   * by a few pixels.
+   * TODO fix that^ in src/components/Tooltip. Should resize(?) the positioning
+   * wrapper if it's close to the edge of the screen. */
+  overflow-x: hidden;
+}
