+++ src/components/SidePanels/index.js
@@ -3,7 +3,7 @@ import { translate } from 'react-i18next';
 import compose from 'recompose/compose';
 import pure from 'recompose/pure';
 
-import Chat from '../../containers/Chat';
+import Chat from '../Chat';
 import RoomUserList from '../../containers/RoomUserList';
 import WaitList from '../../containers/WaitList';
